[spark] Remove SparkCatalogBase
